Implement filtering of disks in FreeBSD
